requirements/dev.txt
-r base.txt
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
argh==0.26.2              # via watchdog
argparse==1.4.0           # via codecov
attrs==16.0.0             # via service-identity
beautifulsoup4==4.4.1     # via feedfinder2, newspaper3k
billiard==3.3.0.23        # via celery
celery==3.1.23
cffi==1.7.0               # via cryptography
click==6.6                # via pip-tools
codecov==2.0.5
colorama==0.3.7           # via pytest-watch
coverage==4.1             # via codecov, pytest-cov, pytest-testmon
cryptography==1.4         # via pyopenssl
cssselect==0.9.2          # via newspaper3k, parsel, scrapy
decorator==4.0.10         # via ipython, traitlets
dj-database-url==0.4.1
django-appconf==1.0.2     # via django-fsm-log
django-configurations==1.0
django-extensions==1.6.7
django-fsm-admin==1.2.3
django-fsm-log==1.3.0
django-fsm==2.4.0         # via django-fsm-admin, django-fsm-log
django-fsm==2.4.0         # via django-fsm-admin, django-fsm-log
django==1.9.7
djangorestframework==3.3.3
docopt==0.6.2             # via pytest-watch
enum34==1.1.6
envdir==0.7
fake-useragent==0.0.8
feedfinder2==0.0.4        # via newspaper3k
feedparser==5.2.1         # via newspaper3k
first==2.0.1              # via pip-tools
idna==2.1                 # via cryptography, tldextract
ipdb==0.10.1
ipython-genutils==0.1.0   # via traitlets
ipython==5.0.0            # via ipdb
jieba3k==0.35.1           # via newspaper3k
kombu==3.0.35             # via celery
lxml==3.6.0               # via newspaper3k, parsel, scrapy
mccabe==0.5.0             # via pylama
newspaper3k==0.1.7
nltk==3.2.1               # via newspaper3k
parsel==1.0.2             # via scrapy
pathtools==0.1.2          # via watchdog
pep257==0.7.0             # via pylama
pep8==1.7.0               # via pylama
pexpect==4.2.0            # via ipython
pickleshare==0.7.2        # via ipython
Pillow==3.3.0             # via newspaper3k
pip-tools==1.7.0
prompt-toolkit==1.0.3     # via ipython
psycopg2==2.6.2
ptyprocess==0.5.1         # via pexpect
py==1.4.31                # via pytest, pytest-capturelog
pyasn1-modules==0.0.8     # via service-identity
pyasn1==0.1.9             # via cryptography, pyasn1-modules, service-identity
pycparser==2.14           # via cffi
PyDispatcher==2.0.5       # via scrapy
pyflakes==1.2.3           # via pylama
pygments==2.1.3           # via ipython
pylama==7.0.9
pyOpenSSL==16.0.0         # via scrapy, service-identity
pytest-capturelog==0.7
pytest-cov==2.3.0
pytest-django==2.9.1
pytest-testmon==0.8.2
pytest-watch==4.1.0
pytest==2.9.2             # via pytest-cov, pytest-django, pytest-testmon, pytest-watch
python-dateutil==2.5.3    # via newspaper3k
pytz==2016.4              # via celery
PyYAML==3.11              # via newspaper3k, watchdog
queuelib==1.4.2           # via scrapy
requests-file==1.4        # via tldextract
requests==2.10.0          # via codecov, feedfinder2, newspaper3k, requests-file, tldextract
scrapy==1.1.0
service-identity==16.0.0  # via scrapy
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via cryptography, django-extensions, feedfinder2, newspaper3k, parsel, pip-tools, prompt-toolkit, pyopenssl, python-dateutil, requests-file, scrapy, w3lib
tldextract==2.0.1         # via newspaper3k
traitlets==4.2.2          # via ipython
Twisted==16.3.0           # via scrapy
w3lib==1.14.2             # via parsel, scrapy
watchdog==0.8.3           # via pytest-watch
wcwidth==0.1.7            # via prompt-toolkit
zope.interface==4.2.0     # via twisted

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via cryptography, ipdb, ipython, tldextract, zope.interface
